refactor: rename customer resource files.
